moved event tracking into their own class file
