+++ src/EA.Weee.Web/RazorHelpers/WeeeGds.ProgressiveDisclosure.cs
@@ -1,14 +1,25 @@
 ﻿namespace EA.Weee.Web.RazorHelpers
 {
-    using System.Web.Mvc;
+    using System;
 
-    public partial class WeeeGds<TModel>
+    public class ProgressiveDisclosure<TModel> : IDisposable
     {
-        public MvcHtmlString CreateProgressiveDisclosure(string linkText, string hiddenContent, string eventCategory, string eventAction, string eventLabel)
+        private readonly WeeeGds<TModel> gdsHelper;
+
+        public ProgressiveDisclosure(WeeeGds<TModel> gdsHelper, string linkText)
         {
-            return new MvcHtmlString(string.Format(
-                @"<details class=""govuk-details"" aria-live=""polite"" role=""group""><summary class=""govuk-details__summary""><span class=""govuk-details__summary-text"">{1}<span class=""hidden-for-screen-reader"">This is an expandable link that will reveal more content upon activation</span></span></summary><div class=""govuk-details__text""id=""details - content"" aria-hidden=""true""><div class=""govuk - body govuk - !-font - size - 16"">{2}</div></div></details>",
-                EventTrackingFunction(eventCategory, eventAction, eventLabel), linkText, hiddenContent));
+            this.gdsHelper = gdsHelper;
+
+            var html = string.Format(
+                @"<details class=""govuk-details"" aria-live=""polite"" role=""group""><summary class=""govuk-details__summary""><span class=""govuk-details__summary-text"">{1}<span class=""hidden-for-screen-reader"">This is an expandable link that will reveal more content upon activation</span></span></summary><div class=""govuk-details__text"">",
+                gdsHelper.EventTrackingFunction("Hidden content", "Progressive disclosure", linkText), linkText);
+
+            gdsHelper.HtmlHelper.ViewContext.Writer.Write(html);
+        }
+
+        public void Dispose()
+        {
+            gdsHelper.HtmlHelper.ViewContext.Writer.Write("</div></details>");
         }
     }
 }
